+++ doc/release-notes.md
@@ -125,6 +125,9 @@ Wallet
 Low-level changes
 =================
 
+- The RPC gettransaction, listtransactions and listsinceblock responses now also
+includes the height of the block that contains the wallet transaction, if any.
+
 Tests
 -----